craft-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pillow==11.0.0
numpy==2.1.0
craft-text-detector==0.4.3
